DS-5284-fixes
fixes for visualisations in table header. Percentage threshold is moved from 70 to 60, size of visualisations is decreased, font of percentage values is changed, colours are properly named, so everything works good when theme is switched and describe command processing is refactored  to correctly avoid possible warnings in its text for pandas and polars, also, added bars to histograms with height 0 to get rid of empty space.


Merge-request: IJ-MR-117579
Merged-by: Georgii Zorabov <georgii.zorabov@jetbrains.com>

GitOrigin-RevId: cf9e75fa51d827266c5b63116769bc662cc0d26b