IDEA-336046 Fix speed search typing with Got It tooltip
Second attempt. Capturing global key typed events didn't
work out because it failed to capture things like Backspace
and because it was too invasive, as the Got It tool tip may
stick there for quite a while (it doesn't close on focus close).

This time do it by not focusing the Got It tooltip, but instead
closing it explicitly when Esc, Enter or Space are pressed,
or rather released, as this event comes last, and we want
to block previous events to avoid passing them anywhere
until the Got It tooltip is closed.

The Got It tool tip API had to be modified to be able to hide
the popup with ok = true, as otherwise it would keep popping
up again and again, as if the Got It button was never pressed.

(cherry picked from commit 5c36557006e6de5fce6f2a8d5b3e9cdab814cba7)

IJ-CR-117688

GitOrigin-RevId: c3222df34cb5757d44614672de92e274e1d3a3bd