[lang] Fix name suggestion race condition when setting initial selection
Before, the selection was done at some indeterminate point in time because of `SwingUtilities#invokeLater` this made it possible for the selection to be done before the dialog was shown. If this happens, the editor is not initialised yet and the selection can't be executed. By using a listener we know for sure that the dialog is shown when executing the selection. #IDEA-335018 Fixed

GitOrigin-RevId: 19d46f594715f24638fcb8846ccd5af9722d45a8