IJPL-310 Refactor processWithPluginDescriptor method for improved sorting control
This update refactors the processWithPluginDescriptor method in the ExtensionPointImpl.kt file and other related usages. The method is now divided into two: the original method, for sorted calls, and a new one, processUnsortedWithPluginDescriptor, specifically for unsorted calls. The aim is to enhance the functionality and control of sorting processes in the extension handling.

GitOrigin-RevId: 0a518645c6dda874f77b692088243a3ab429e717