sh: suggest to run shell scripts from physical files only (IDEA-336310)
Even though running shell script not from physical files is possible (for example, statements could be run in the project root folder, or in the parent folder of containing file - relevant for shell code fences in *.md files), it seems like `ShRunConfigurationProducer` doesn't support it yet, because it sets `configuration.setScriptPath(virtualFile.getPath())`. If such a feature is wanted, it should be designed separately.

GitOrigin-RevId: 80c3f0d3a61862c8f00159e6e4c42d034ea358f7