IDEA-336046 Don't keep speed search when a Got It tooltip is shown
It was done when the Got It tooltip was focused, to prevent
both the speed search field and the tooltip from disappearing
immediately because the speed search field lost focus.

Now that it no longer loses focus because the tooltip isn't focused
anymore, we don't need that hack either, so everything
can now be hidden on focus loss as usual.

(cherry picked from commit a765bcbfa42183f256672fbcd1e4f16b2e361234)

IJ-CR-117688

GitOrigin-RevId: 760977f1541f9ba662dbabda738dedfb67c4676b