IDEA-230738 Use invalid stored tool window bounds by fitting them into the screen
If the stored external decorator (windowed/floating) bounds are
invalid, they were completely discarded and then the actual
location and size of the window would be used, falling back
to the preferred size and centering the tool window relative
to the IDE frame. Apparently that didn't work well because
at this stage the preferred size can be too small for some reason,
possibly because the tool window's contents may not be initialized
yet.

So what we try to do instead now is that if the saved bounds
aren't empty, we still try to use them by moving to fit into the screen,
cropping if they don't fit, and then centering relative to the IDE frame.

While this may cause the window to become too large if the new
screen is too small, it's easier for the user to resize a large window
(as long as it fits into the screen) than to resize a window that's
incredibly small.

GitOrigin-RevId: 0d097e27d5cf885a2eb98ebdb0b8c7b4809714a4