IDEA-335458 Fix race condition while preparing onboarding tips
Now prepareOnboardingTips activities set a sign to install onboarding tips on the first opened editor. So it may be (and should) executed before any files will be generated and opened by project wizard.

Also wanted to rename prepareTipsInEditor to prepareOnboardingTips (because installing to the editor will be later by the fact), but cannot do it because of API break.

GitOrigin-RevId: 1fdff5f9d99cffebb2a525f82374763be86ef4fb