[javadoc] Fix relative URLs conversion in JavaDocExternalFilter
When the `href` attribute of an anchor tag starts with a single forward
slash ('/'), this means the link should be relative to the authority
(host + port) of the current page. Similarly, when the `href` attribute
of an anchor tag starts with two forward slashes ('//'), the link should
be relative to the scheme (e.g. https://) of the current page.

closes https://github.com/JetBrains/intellij-community/pull/2597

GitOrigin-RevId: a8888f2a5dc283d158596229d6d602e1a2f8d265

(cherry picked from commit bf0c93b0a51ee4393e2333998f2d69f784b280c0)

Change-Id: Ibf0c93b0a51ee4393e2333998f2d69f784b280c0
Fixes: 308175144