This change addresses https://youtrack.jetbrains.com/issue/IDEA-336443
When the `href` attribute of an anchor tag starts with a single forward
slash ('/'), this means the link should be relative to the authority
(host + port) of the current page. Similarly, when the `href` attribute
of an anchor tag starts with two forward slashes ('//'), the link should
be relative to the scheme (e.g. https://) of the current page.