[javadoc] Fix relative URLs conversion in JavaDocExternalFilter
When the `href` attribute of an anchor tag starts with a single forward
slash ('/'), this means the link should be relative to the authority
(host + port) of the current page. Similarly, when the `href` attribute
of an anchor tag starts with two forward slashes ('//'), the link should
be relative to the scheme (e.g. https://) of the current page.

closes https://github.com/JetBrains/intellij-community/pull/2597

#IDEA-336443 Fixed

GitOrigin-RevId: a8888f2a5dc283d158596229d6d602e1a2f8d265