IDEA-250563 Allow some tolerance when fitting IDE frame into the screen
The actual bounds of the frame may be slightly outside the screen
for some reason, most likely due to drag zones around the borders.
This causes the bounds to be incorrectly restored if the frame was
touching a screen edge when it was closed.

Fix by adding some tolerance, allowing it to be slightly outside.

(cherry picked from commit e1475bb3b14fdaf0116f1aa5b745ffefa2ef11fc)

IJ-CR-118128

GitOrigin-RevId: 477e2c4205a08b0afee4b42474d0c412f969be03