IDEA-250563 Allow some tolerance when fitting IDE frame into the screen
The actual bounds of the frame may be slightly outside the screen
for some reason, most likely due to drag zones around the borders.
This causes the bounds to be incorrectly restored if the frame was
touching a screen edge when it was closed.

Fix by adding some tolerance, allowing it to be slightly outside.

GitOrigin-RevId: e1475bb3b14fdaf0116f1aa5b745ffefa2ef11fc