[performance-plugin] Improve performance testing wait for smart mode
This commit improves the performance testing wait for smart command by using the `runWhenSmart` method instead of `smartInvokeLater` to ensure that the command execution is delayed until the project is in smart mode. This prevents potential issues when executing the command while the project is not in smart mode state.

^AT-595

GitOrigin-RevId: aa1b73556a9a14f762806b968aa89e49116b9617