IDEA-336686 Fix BgtTreeWalker visiting nodes in reverse order
A very silly bug: instead of removeFirstOrNull(), removeLastOrNull()
was called, probably because of a similarly named lastOrNull()
above (which is correct, because that one is a stack).

This resulted in the walker visiting nodes in reverse order,
which affected pretty much any case where a visitor should
be either collecting nodes in specific order or should find
the first one, and not the last one.

(cherry picked from commit 6817cee4a98a667122f1a8ac0b097cc8b9b5cdfe)

IJ-CR-118513

GitOrigin-RevId: 2ecb75cd938ac22e46ebfc197bac15065de4b500