IDEA-333308 Disable multi-selection in the Project and Packages Views
For many projects, for example Gradle ones, it causes a lot of
external library nodes to be expanded just to select multiple
occurrences of a library class. This is weird UX and causes
performance issues.

Since multi-selection was added on purpose, it may be that
some specific project view panes may actually need that.
So keep the default behavior in AbstractProjectViewPaneWithAsyncSupport,
but override it for the Project and Packages views.

GitOrigin-RevId: c82438a6cc36b5a028210aa40a571636bb56e889