IDEA-336417 Update nodes in the Scope Chooser
Node update was introduced along with AsyncTreeModel
in the Package Dependencies, but since the Scope Chooser
uses the same tree model, it led to a regression.

Ideally the Scope Chooser should be migrated to AsyncTreeModel
as well, but since it's a regression we need a quick and safe fix now.
So just update the nodes when they're sorted, as it seems a convenient
place to put these updates. They, of course, involve slow ops,
but those should be fixed as a separate effort, likely when migrating
to AsyncTreeModel.

(cherry picked from commit 4c9428b5688261f6011174b1c824a6402f5d85d2)

IJ-CR-118514

GitOrigin-RevId: 1d2c963740df1b6e252d2dc9eafd0abec0fd1488