IDEA-336917 Refactoring: rename stuff in StructureViewSelectVisitorState
The term "length" here was confusing: it could refer to either the path's length
(which it does) or the element's length (which it doesn't). We're about to introduce
the element's length into this mess, and it won't look good without this refactoring.

(cherry picked from commit 59145bff30cae4e18d1fcbafe6eb67ee9ef4390d)

IJ-CR-118513

GitOrigin-RevId: e0314b3555e3c6b07199674e69eed93ab860a6c3