IDEA-336917 Use the shortest match in the structure view
The existing implementation just picked the first of the elements
having the same depth in the tree. This decision was arbitrary
and in theory there should have been only one deepest element
that is matching the search criteria (either the same element
as the one we're looking for or its ancestor).

However, when the PSI and the tree don't match (as is the case
with Ruby), it's sometimes possible that the ancestor and the element
are placed on the same level. In this case, the shortest match
should be preferred, which is exactly what's fixed here.

(cherry picked from commit d638247f9bd248f89eb87ffa990c38594c44a4e8)

IJ-CR-118513

GitOrigin-RevId: 7ef6640ca94c39ced67d37c68d1dbcdc249d37b2