IDEA-336991 Fix Timer leak in SpeedSearchBase
For some reason, the text field wouldn't stop the Timer
responsible for the caret blinking even when the field
is hidden, which would cause the Timer to remain on
the global event queue's listener list, causing plenty
of leaks.

Fix by making the caret invisible explicitly in removeNotify().
Also, just in case, add isShowing() check to the focus
event handler (which is never invoked anyway,
but better be safe than sorry).

GitOrigin-RevId: e6f7095e6fdadda48555856f4d94f859a0dfec02