ML in SE: Expose and modify EvaluationResult to use as a key (IDEA-334557)
Our logic in the ML in SE plugin, relies on unique IDs
for every element found. It is basically an Any->Int map.
We did not have a key provider for CalculatorSEContributor,
so we are going to expose it.

Also, to ensure that two evaluations that have the same result,
but different expression, we are going to include the expression
in the EvaluationResult.

GitOrigin-RevId: 4cdfb6ac7a17d6b5cecec3f5b767b0da90d0ac01