[Workspace model] IDEA-334293 Fix test name `EntityStorageSerializationTest.immutable serializer version prefix`
The test is needed to check that the cache version prefix is immutable.

After adding metadata to the caches (IDEA-300053), it was decided to reset the cache version. After resetting some users had the same version in the old cache as in the new one. Since the cache structure has changed, error was thrown.

To correct this it was decided to use a new version prefix in the cache. To avoid such errors in the future, this test was added.

(cherry picked from commit c62d5b64139e5ef9faa0134b6f9396427f147835)

IJ-CR-117927

GitOrigin-RevId: baecb0b79fc139337183534a95a9e865931d420f