notifications: show tool window notification if tool window bars are hidden: IDEA-336977
If the tool window bar is hidden, we can't detect where would the toolwindow button be located. Thus - the same as for the old UI notifications - let's show the notification for the component, which is the toolwindow itself.

Let's also remove the balloon pointer, because it would point to nowhere (especially in the new UI, where it points either to the left, or to the right, while the toolwindow can be located at the bottom).

GitOrigin-RevId: b793d9e1d63ebb9709a239a8c07bff7f26be5314