[cross-platform launcher] allow providing custom launch data for some commands in product-info.json
To run JetBrains Client from installation of a big IDE (RDCT-521), we need to use different classpath, main class, vm options and data directory name. This commit adds support for `customCommands` property to product-info.json and supports it in the cross-platform launcher. If the command line argument is present in `commands` array, the corresponding object is used to override values from 'launch' object.

GitOrigin-RevId: 2801ab568324936173f7408a01c48713cebd9b47