[java rd] use fake root packages for modules included in com.intellij.java.frontend plugin (IDEA-322563)
We need to have root packages for module xml files in plugin model v2, but if we specify packages with real classes in these modules, they'll fail to load with "must not be requested from main classloader" error, because inter-module dependencies aren't specified for these files. If we specify module dependencies, the big 'com.intellij.java' plugin will fail to load because it isn't migrated to plugin model v2 yet and includes contents of these module descriptors via xi:include tag. So it seems that we need to use fake packages without real classes until 'com.intellij.java' plugin is migrated to plugin model v2.

Check in InterfaceExtensionImplementationClassResolver is suppressed for 'com.intellij.java.frontend' plugin as well for now.

GitOrigin-RevId: 7b2321835becc9f4589ddae09a40160bf735044e