[platform] support loading plugins with multiple classes roots using the module loader (RDCT-688)
When JetBrains Client is started from IntelliJ IDEA installation, it loads some modules of 'com.intellij.java.frontend' plugin from Java plugin's lib directory, and the main module is loaded from a separate JAR. So in order to support that, ModuleBasedProductLoadingStrategy is changed to handle plugins with multiple resource roots. ModuleBasedPluginXmlPathResolver is added to support loading of module descriptors from other JARs.

Also, problems with loading plugins in ModuleBasedProductLoadingStrategy are reported as warnings to the log and doesn't break startup.

GitOrigin-RevId: 8193f6dd100868e5757a54d4e615a7261cf636d0