KTIJ-27596 [scripts]: highlighting issues in .gradle.kts
What was the problem caused highlighting issues: missing script definition.

The class `ScriptDefinitionsManager` has implicit contract: `definitions` are either null or non-empty list.
There are statements like definitions?: <do-smth>. The contract was broken.

Why was a definition missing?

There is `ScriptDefinitionsManager.reloadDefinitionsBy(<source>)` which can be called as a reaction to some event.
When a user opens groovy files (`build.gradle` for example), the method is called as a reaction to the compiler
settings update with the corresponding source. It loads nothing and leaves definitions empty (not null).
Further calls to `ScriptDefinitionsManager.findDefinition()` (the user opens, say, `build.gradle.kts`) bumped
into `definitions?: <load>` and result in nothing.

^KTIJ-27596 fixed

GitOrigin-RevId: 0fdadde19195e42a8ee7f841c02dec54a09dfd07