[groovy] IDEA-260789 IDEA-244230: Support custom VM options in greclipse
Modifying VM options is not allowed in the process itself, so we have to start a new Java process.

Also, Maven provides compiler argument 'javaAgentClass' to greclipse, which is not well-formed VM option. We have to workaround it by custom recognition of this particular argument.

Also, this commit rewrites configuration panel to a modern UI DSL

GitOrigin-RevId: 67da035ba224434026d0628490b6811687f684be