IDEA-332385 Don't process KEY_TYPED events with modifiers in Speed Search
What happens here is when Ctrl+N / Ctrl+P is pressed, it generates
two events: KEY_PRESSED and KEY_TYPED. The first one is processed
correctly, and it does select the next / previous element. But then the
next event arrives and screws up everything because the Speed Search
"thinks" it's actually processed somehow (like if an actual character was
typed), but fails miserably because nothing has really happened.
This eventually leads to the selection being updated incorrectly
(as if something really changed, but it didn't).

Fix by simply filtering out such typed-but-not-really events and ignore
them without consuming (this is important because the speed search
processes everything that's consumed by the field).

GitOrigin-RevId: be1fe0d3cbec82fced1afe5697c25c2d4d2906dd