IDEA-307800 Consume navigation key presses in SpeedSearchBase.SearchField
If it's not done, then SearchPopup.processKeyEvent ignores these.
Standard shortcuts are consumed by the field itself (in the super call),
but if we use custom ones, like Ctrl+P / Ctrl+N, then the field has
no idea about those and doesn't consume them, which causes the
speed search to ignore them.

So if an event wasn't consumed, but it's known to be a navigation
shortcut, fix this behavior by consuming it explicitly.

GitOrigin-RevId: 3850bc5c3cc9faf7a8aac18ceff9e1fba8a83f86