IDEA-65709 Fix custom navigation shortcuts not working with speed search
This is a weird one. As it stands, in the current implementation
this issue is caused by code that fixed (somewhat more recent)
IDEA-149094. It explicitly disabled scrolling actions as long
as the speed search is active. However, in the current implementation
such behavior doesn't make much sense because the same code
(ScrollingUtil._move) was invoked through the alternative
actions, namely ListMoveUpAction and similar. The only difference
is that those alternative actions are Swing actions (as opposed
to IJ AnAction) and therefore don't recognize IJ shortcuts.

Fix by enabling scrolling actions again, which looks like a harmless
thing to do now. Whatever they were trying to fix is probably
implemented in SpeedSearchActionBase now that has its own
key event handling.

GitOrigin-RevId: 34dafa41149e4f92cba4279221eae371930686e4