[Kotlin] Properly handle extension receiver parameter in debugger without a label
Such label's handling is more inlined with `StackFrameProxyImpl` and how
it stores variables. In K1 we will be able to find parameter right away,
but this is not the case for K2. For some reason, the naming of variables
in `StackFrameProxyImpl` is different for K1 and K2. We are going to use
unsafe fallback. For that reason, some adjustments in `VariableFinder`
were required.

#IDEA-335204 Fixed

GitOrigin-RevId: 8111b65e3e807b224e10d7470f16f7235b76b3c1