KT-61890 [kotlin] Adjust testdata for new scope provider implementation
There are multiple potential problems:
1. Due to more correct scopes in the supertype/constructors positions,
`InSecondaryConstructor*` tests have diverged from the K1 behaviour -
it's no longer possible to see inaccessible class members at the
constructor's default parameters positions. See KTIJ-27753
2. `PreferMoreSpecificExtensionLocalClass.kt` stopped working because
there is a variable `c` which references the old local `C` class
instead of the new "on-air-resolved" `C` class.
This, in turn, breaks `KtCompletionCandidateChecker`, and the
shadowing in completion.
Similar thing with `WhenWithSubject.kt` test - there is an
old symbol for `some` variable which is attached to the PSI,
and it's different from the `some` symbol obtained from the
scopes. See KT-63320

GitOrigin-RevId: 395ad5dcc9005adb6478c1817143726ab009396f