IDEA-337560 Add DisposableWindow interface and use it to skip disposed windows in uiTraverser
Unfortunately, Swing provides no way to figure out whether a window was disposed,
and Window.getWindows() keeps returning disposed windows until they're garbage
collected. This causes all sorts of issues when traversing the whole UI, for example,
in ComponentTreeEventDispatcher, that invokes callbacks like uiSettingsChanged
on every component implementing the corresponding interface. When such a component
is located in a disposed window, all kinds of things can go wrong.

Note that this isn't technically a leak, as Window.getWindows() list uses weak references
internally, so the windows can still be garbage collected. The issue is that callbacks
on those components can access all sorts of already disposed resources and break
in a rather unexpected way.

Fix by introducing the DisposableWindow interface which we'll now have to implement
in every window we care about.

Implementation note: the way the UI traverser is currently implemented, this
new interface will allow us to skip all the children of a disposed window, but
not the window itself. This may have to be fixed separately if it ever becomes an issue.

GitOrigin-RevId: a7d26bfff47c9ddb8580410fb96199571e190557