IDEA-336525 IJent: prevent file descriptor and memory leaking in processes
Two different leaks were found and fixed:
* Memory leak in a hash table inside IjentProcessesService. It used to populate but was never cleaned. Actually, this memory leak is negligible.
* A leak of a file descriptor associated with the stdin of a child process. This leak brought troubles.

The file descriptor of stdin was held until the corresponding gRPC stream is closed. The gRPC stream is bound to the byte channel of IjentChildProcess in the IDE. The byte channel is bound to custom implementations of java.lang.Process.

It's a usual pattern to never close pipes of java.lang.Process, because it rarely brings any benefit. There are a lot of internal buffers, and java.lang.Process closes file descriptors as soon as possible by itself, not waiting until the corresponding InputStream and OutputStreams are closed. This implicit contract was broken, leading to file descriptor leakage.

Since this commit, IJent closes the gRPC stream for stdin as soon as the process dies. It shouldn't break anything on the IDE side, because an attempt to write to that stream anything would have caused an error anyway.

GitOrigin-RevId: f7a625cfa5f9b6080f9253360fc5eb52c70b9e47