IDEA-337630 Account for border width in completion popup text indent
Since the popup panel has a border, at least in the new UI,
it needs to be accounted for when we calculate the offset to
shift the popup by when aligning it with the text in the editor.
Otherwise, the popup appears to be slightly shifted to the right.

(cherry picked from commit b67bd00ad007f25e8c5611263a9b379a37c42a08)

IJ-CR-119525

GitOrigin-RevId: 5a026e0501e102f6f54bc7fb5d638759a5f6dfe0