IDEA-337630 Implement empty icon customizations for LookupCellRenderer
The icon customizations done by AI suggestions and ML ranking
were screwing up the popup's alignment because the empty
icon width is used in getTextIndent().

Fix by adding a new method to ItemPresentationCustomizer,
customizeEmptyIcon(), that is invoked to compute an empty icon
that would account for customizations.

Implementation notes

- An updateIconWidth() call is added to addPresentationCustomizer(),
because we need to recompute the empty icon's size even if nothing
else has changed.

- The updateIconWidth() implementation is now less efficient because
it needs to customize the new icon before comparing it with the
existing one, as that one is already customized.

- The branch that would un-decorate the icon in augmentIcon()
had to be removed because it was only needed to compare its
size with the empty icon (passed as "standard"), which is now also decorated.

(cherry picked from commit 5c4982af780a37d7d8eb41868e0f0adc871cc45d)

IJ-CR-119525

GitOrigin-RevId: a66aed2d0a281ce900c0c0c911f69e7bb18bfdae