GTW-6821 Fix Git widget popup is misplaced on a second monitor
Abstract popup adjust a popup position, so a popup would fit the screen based on the component coordinates (`fitXToComponentScreen`)

In RD component coordinates might be (0,0), so AbstractPopup will try to move the popup to the first monitor.

In `CwmPopupInterceptor` we process such popups and adjust RelativePoint, so it'll provide the proper position, though we don't alter the component itself.
Instead of altering passed RelativePoint, we can explicitly intercept the popup and show it in the screen coordinates RD provided

GitOrigin-RevId: e011a7d5f10c01e1506ba347a3c04618ec650d3a