IDEA-337817 Add some tolerance to maximizing/normalizing dialogs
Since we're using a hack to maximize/normalize dialogs (as they
don't support it natively), when the dialog is "maximized",
it's still movable. This creates a possibility to move it accidentally
by a few pixels when trying to double-click the header to normalize
the dialog. This results in the dialog maximizing again and forgetting
its normal bounds that were saved previously.

Fix by introducing some tolerance, so trying to normalize a dialog
that's "almost" maximized will still normalize it properly.

GitOrigin-RevId: 63004e658fc24b232b29182611ddf82ce7835d85