[Workspace Model] Get rid of outdated functionality of sending broken storage with the exception
We stopped sending the storage itself in e875200e75daace4fd50fd60098a5a52d4218173 ([Workspace Model] IDEA-335463 Remove the diff log and storage serialization in the `reportErrorAndAttachStorage` method) when there were issues with the metadata.
However, we still had related classes and logic.
With this update I clean up remained things.

This should improve the performance with cases where we check for consistency during addDiff operation. Because now we don't create a snapshot of builder (which is not used anyway).

This change is also needed to get rid of JpsProjectSerializersImpl.squash function and use the regular add diff. Because, even if getting rid of squash function will make IJ faster, the conversion to the snapshot on every addDiff operation will make tests significantly slower.

GitOrigin-RevId: 6af527637cfedfb6b3e3985243ee586e7abb1617