IDEA-337630 Recompute empty icon width on LookupUi.refreshUi
When, say, ItemsDecoratorInitializer.markAsReordered is called,
the state of the customizer may change in such a way that
the empty icon width should be updated as well.

Unfortunately, this happens after the icon was customized,
but fortunately, before the text indent is actually calculated.
So we need to update it. Do it explicitly by adding the refreshUi
method to LookupCellRenderer and call it whenever LookupUi.refreshUi
is called to update the width.

(cherry picked from commit 9adfc9f72086ebf150caa522e04ba592a47f59f5)

IJ-CR-119525

GitOrigin-RevId: fb44717294c102fc34f014234badd114a28ce4e3