IDEA-338005 call `indicator.start` after the progress indicator is shown
Calling `indicator.start` before adding in the UI resulted in `InlineProgressIndicator.wasStarted` remaining `false`,
which is turn didn't display "Stopping - " prefix for cancelled coroutine progresses.
`indicator.start` should be invoked _after_ the progress is shown in the UI, because when it's added to the UI
`InfoAndProgressPanel` registers a bunch of state delegates, one of which tracks `indicator.start`.

30f40920 was originally intended to fix the problem, but had to be reverted because of IDEA-327037.

GitOrigin-RevId: d2576dd2d846daa0b022e91991660fee85f03280