IDEA-338097 Un-dispose IdeFrameImpl on show()
Turns out we reuse the disposed frame when entering full screen.
This wasn't accounted for when the DisposableWindow interface
was implemented, so upon entering full screen the frame was
no longer being traversed by uiTraverser, which caused, for example,
UI resource colors to not update properly, as they're updated
when updateUI() is called for all components by the LaF manager.
This, of course, broke much more than just colors, those were just
the most obvious thing.

Fix by simply resetting isDisposed to false when the frame is shown.

GitOrigin-RevId: d17f083f5dfcb1ce7b67dd9bdc66edea725089c2