IDEA-338005 call `indicator.start` after the progress indicator is shown
Calling `indicator.start` before adding in the UI resulted in `InlineProgressIndicator.wasStarted` remaining `false`,
which is turn didn't display "Stopping - " prefix for cancelled coroutine progresses.
`indicator.start` should be invoked _after_ the progress is shown in the UI, because when it's added to the UI
`InfoAndProgressPanel` registers a bunch of state delegates, one of which tracks `indicator.start`.

Comparing with d2576dd2, the current approach uses proper scope nesting
making sure `finally` block is executed if `showIndicatorInUI` completed.

GitOrigin-RevId: fce065438892b969aaa798798f20245c9a7d5429