honor ActionRemoteBehavior spec in update session
Fixes cache-misses in an editor context menu in "RevealGroup". `RevealFileAction` is enabled on server but disabled on the client side, so the group tries to access the next child to check if it is enabled. At the same time, in CWM mode, the action must be enabled for a host user.

GitOrigin-RevId: 9d9ff3adf9b8b98fcf9400a3802e22acf59be18f