IDEA-337980 IJent: rewrite process communication code to get rid of races
This huge commit fixes several problems caused by the former flaw in the API.

The old API implied that `start_process` returns immediately as soon as the process is launched. When API users needed to read stdout/stderr, write to stdin or wait for the process exit, they should send separate requests providing the process ID.

Such design led to existence of a global hash map, where the key was the pid, and the value was the smart pointer to a mutex with a process.

The first problem was that the hash map was never cleaned: IDEA-336525 It led to memory and file descriptor leakage. For example, a test run of PyCharm + WSL Interpreter make the whole IDE unusable just in a couple of minutes, because all file descriptors became occupied.

A clumsy introduction of the method `close` revealed more awkward problems:
* After calling `close` the process is removed from the global mapping. Hence, all consequent requests to API led to an error that the process isn't found. The biggest problem was that `wait` was organised as a long polling command. This race used to happen very frequently:
    * The process started.
    * A background thread was calling `wait` in a loop.
    * `wait` worked for 3 seconds and returned that the process is still alive.
    * The process terminates.
    * `AutoClosingIjentChildProcess` detects the process death and calls `close`.
    * The loop executes `wait` again and fails with an error.
* The races like above also used to happen with `stdin`. It actually was a bug inside `AutoClosingIjentChildProcess`, but that class wouldn't have existed if the API had been done right.
* The process was stored as `Arc<Mutex<Process>>`. The API of underlying `Process` dictates `wait` to borrow the reference to the process until the end. Hence, the whole mutex was holding with `wait`. Luckily, there was a timeout for 3 seconds, but anyway it made impossible to terminate, kill the process or resize its PTY immediately.

The new API creates a stream of events from the process. The termination of the stream implies the death of the process. Since access to stdin, stdout, stderr and the exit code doesn't require access to the hash map anymore, the race disappears. Also, the mutex disappears too, because the streaming coroutine becomes the only owner of the process.

The hash map still exists, though it holds weak references to the processes. It implies no file descriptor leakage, at least through the map. API calls for process termination and PTY resizes still work through the map.

It's important to keep the terminate/kill API calls separate from that event stream, because it's crucial to be able to send those requests immediately. Delivery of such requests through the queue shared with stdin data chunks might have lead to inability to kill the process at all, if the process is stuck and doesn't read stdin.

GitOrigin-RevId: a9fc3b9a0c7121aac16da1404b8f6c4ce9598d2e