IDEA-338240 Update root pane on main menu button hierarchy events
The leak is caused by the root pane being temporarily set to null
when the frame is released to reuse for another project.
This happens in WindowManagerImpl.releaseFrame and causes
a hierarchy event to trigger. However, during this event
the button is still considered to be showing, which is why
the uninstall() call never happens. Then a new button is created
for the new project and the old one is leaked.

Fix by explicitly updating the root pane on every hierarchy change
event. In the worst case, when the root pane doesn't really change,
it's a no-op because the setter compares the new value to the old one.
However, if the root pane does change, even if to a null value,
then the setter triggers uninstall() properly.

Add a null check to the setter as well to avoid a log warning
in install() when that happens.

(cherry picked from commit 9b887d17efeff03f1f30c85960d18fc92bd0c939)

IJ-CR-120123

GitOrigin-RevId: 2fefd3ed8d081bf0bdc554ea88637cc813d684b2