IDEA-338293 Catch exceptions thrown by globalSchemeChange
This interface can be implemented anywhere, including 3rd party
plugins, and we don't want some plugin to throw an exception
and break our code somewhere in the middle of a LaF change,
which causes the platform to end up in a weird state.

So catch these and log them right in EditorColorsManagerImpl.

GitOrigin-RevId: c32da4f9f0698f7a4e8c607f21552cde7415a73a