IDEA-337792 Skip floating toolbars in ToggleToolbarAction
This action should affect only permanently visible toolbars.
It didn't really toggle the visibility of floating toolbars,
as they manage it themselves, but it was picking up actions
from those toolbars when this option is off. This led to weird
actions appearing in the tool window menu out of the blue.

Fix by simply skipping all floating toolbars when traversing
toolbars.

GitOrigin-RevId: 7ad712ceab346a70513292ce45a7c7153a8d4caa