treat all groups as dynamic in action customization
Fixes missing VCS Widget in rd client. The widget is present on server ("MainToolbarVCSGroup"), but the client has its own customized `DefaultActionGroup` ("MainToolbarLeft") with actionId that is not queried in update session because of mod-stamps check. In the old update logic, that works because actions are really inserted into groups.

GitOrigin-RevId: ceb6419d6bf8e6ac88150df1ce904275552905ab