IDEA-329731 Fix absence of New UI Onboarding if New UI was enabled from Welcome screen banner
The problem is that we store the `experimental.ui.on.first.startup` property only when a project is opened. So, there can be the case when user starts the IDE, Welcome Screen is showing because there are no opened projects, user enables the New UI from the Welcome Screen banner, restart, and open the project. And there is no New UI Onboarding proposal because we think that New UI was already enabled on the first startup, because it is the first time when project is opened, and `experimental.ui.on.first.startup` property is stored.
Changed to store `experimental.ui.on.first.startup` additionally when the Welcome Screen is shown.

GitOrigin-RevId: c44f6a42644b902266b4da506e6dd04134b3c06a