[terminal] IDEA-338348 Disable output caret model when alternate buffer is enabled
Alternate buffer view has its own caret model and painter, so when it is enabled, the output caret model should be disabled, to not clash with the alternate buffer caret model. Also, there is no need to repaint the caret in the output editor at this moment, because it is not showing.

GitOrigin-RevId: 0232ccd2ce830f022099ce23c9538cbc1983ea6f