WI-75139: indexingCharLength in IndexingStampInfo should be stored without precision reduction
in stub-consistency checks we actually use indexingCharLength, not indexingBytesLength, so indexingCharLength should be stored in full precision (31 bit). At the moment, IndexingStampInfoStorage capacity is 4*int per virtual file (due to page size adjustments, stored element size can be 2 or 4 ints), so there is 1 unused int per file in the store (old storage format only needed 3 ints), which we will use to fix the problem: indexingCharLength should not be reduced.

GitOrigin-RevId: bf42dbbf2dd9c12f46a1b626222893ab76ce6ef3