[external system] ExternalSystemProcessingManager: do not update ended progress in onStatusChange/onTaskOutput
In some rare cases clients can report `onTaskOutput` even after `onEnd` is already called.
This might happen when some external system process first produces some "end" event but then still prints some output to the STDOUT.

This might lead to an issue that `myTasksInProgress` map is forever non-empty
(because event.getId is unique and onEnd is already invoked)
As a result `ExternalSystemProcessingManager.hasTaskOfTypeInProgress` will always return true.
Due to this "Reload All Projects" action can get stuck forever until you restart IDE

See details in related Scala Plugin issue:
https://youtrack.jetbrains.com/issue/SCL-21794/Reload-ALL-BSP-projects-action-is-disabled#focus=Comments-27-8367670.0-0

It's not 100% clear whether it's a client responsibility to not call `onTaskOutput` after `onEnd` is called.
And it's not 100% clear what to do with such output.
So let's just log a warning for it.

GitOrigin-RevId: 02314801d061eb76f2e0035e21aee81239b41c11