IDEA-338758 Remove imported theme/scheme mapping when storing a new one
2023.2 used the theme name as the key, 2023.3 uses the ID.
What went wrong:

1. Some theme/scheme mapping is stored in 2023.2 using the theme name.
2. The mappings are imported into 2023.3.
3. When the default scheme is selected for the same theme, the ID/scheme
mapping is removed, but the imported name/scheme mapping remains.
4. When the same theme is selected again, the fallback in getPreviousSchemeForLaf
kicks in, selecting the scheme that was stored in 2023.2.

The fix: simply remove the imported mapping when a new one is
stored, whether the new one is to the default scheme or not.

A possible side effect: if there's another theme that has the same
ID as the name of this theme, we'll remove that mapping.
Such a coincidence is unlikely, though.

GitOrigin-RevId: bf5800e128b427e49e9af7d8acf1b9b82d931984