[terminal] IDEA-338730 Show plain terminal in RemoteDev for New Predefined Terminal Session action
On host, it looks strange when the New Predefined session action creates a new Terminal, when a New Terminal Tab action creates a plain terminal. So, it is better to create only plain terminal in this case. Also, because the new terminal is not ready to be shared.
On thin client, it also looks strange, so forcefully create plain terminal.

GitOrigin-RevId: d1bddc6b4a2e0f06eccfd2104121e1e6566954bd